feat: add readline module
